Add support for passing type hints for invocation via generic syntax
This allows the dynamic to properly resolve parameter types when the
parameters are also dynamic or object, and it needs help determining
which overload to invoke (or to match the signature).

Example: dyn.DoSomething<Foo>(obj)

Where the obj may be a COM reference and not trivial to figure out
if it supports Foo type.