Fix support for shared projects
This one was tough to crack:

Several targets replace wholesale the entire @(EmbeddedResource) items, causing the original project
information to be lost and replace with those target files instead, i.e.
https://github.com/Microsoft/msbuild/blob/master/src/Tasks/Microsoft.CSharp.CurrentVersion.targets#L121-L125 and
https://github.com/Microsoft/msbuild/blob/master/src/Tasks/Microsoft.Common.CurrentVersion.targets#L1241-L1244
so you may end up with one or the other targets as the "DefiningProject" for @(EmbeddedResource), depending on
which target run last.

So we instead preserve the original values before the build starts, so we can use our metadata later on.