Make sure we only process embedded resources intended for generation
We found in some cases the @(EmbeddedResource) collection would contain items with the 
`.g.resources` extension which we were not supposed to be processing, and that caused failures. 

Also, the batching was done improperly, which was also causing the issue in the first place. 

So in addition to cleaning up the conditions and batching, also improved the defaulting of the 
item metadata values we expect so they always have a value we can use.